requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.10
Werkzeug==0.16.0
aniso8601==8.0.0
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
marshmallow==3.2.2
python-dateutil==2.8.1
pytz==2019.2
six==1.13.0
flask-swagger
flask_swagger_ui
flask_migrate
psycopg2==2.7.6.1
python-dotenv
nose
